chore: refactor, exploring wasm
having wasm bindgen issues when using audio and input modules.
theoretically `kira` and `midir` can work with WASM, but maybe
not in the way the code is setup right now.

Best option is likely to build things up again from a simpler
project and find the breaking point.

Build small prototype of audio and/or midi