feat: allow refreshing attached midi device
this is an indirect solution to the problem of "get midi working" in WASM,
but it's also generally useful, I think.

It seems like MIDI wasn't connecting before in browser because
the device wasn't immediately available. Trying again in a few seconds,
it works.

I was able to determine this by recreating this polling example locally
https://github.com/Boddlnagg/midir/blob/master/examples/browser/src/lib.rs